src/ingestion/requirements.txt
# ========================================
# Ingestion Pipeline - Requirements
# ========================================
# Install from root: pip install -r requirements.txt
# This file lists only ingestion-specific needs

# Core dependencies (install from root requirements.txt)
python-dotenv>=1.0.0
boto3>=1.29.0
elasticsearch>=8.11.0
requests>=2.31.0
markitdown>=0.1.0
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
langchain-text-splitters>=0.2.0
pydantic>=2.5.0

# Note: All dependencies are in root requirements.txt
# No additional packages needed for ingestion pipeline
